refactor directory monitor
